rename files and folders to match angular conventions
